feat: move game mapping
Criação da rota para chamada do serviço GameListService.moveGame

+++ dslist/src/main/java/com/intensivojavaspring/dslist/dtos/ReplacementDTO.java
@@ -1,3 +1,3 @@
 package com.intensivojavaspring.dslist.dtos;
 
-public record ReplacementDTO(Integer srcIndex, Integer destIndex) {  }
+public record ReplacementDTO(Long srcIndex, Long destIndex) {  }
